refactor: Job 실행 코드를 클래스로 분리
